Authorization with Gates

+++ app/Http/Controllers/Api/AttendeeController.php
@@ -16,6 +16,11 @@ class AttendeeController extends Controller
      */
     use CanLoadRelationShips;
     private array $relations = ['user'];
+
+    public function __construct()
+    {
+        $this->middleware('auth:sanctum')->except(['index', 'show','update']);
+    }
     public function index(Event $event)
     {
         //
@@ -66,9 +71,11 @@ class AttendeeController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $event, Attendee $attendee)
+    public function destroy(Event $event, Attendee $attendee)
     {
         //
+        
+        $this->authorize('delete-attendee', [$event, $attendee]);
         $attendee->delete();
 
         return response(status:204);
